rag_service.py
# rag_service.py
import time
import os
from dotenv import load_dotenv

# Загружаем переменные окружения (например, API токен HF)
load_dotenv()
HUGGINGFACE_TOKEN = os.getenv("HUGGINGFACEHUB_API_TOKEN")

# Импортируем библиотеки, которые ПОТРЕБУЮТСЯ для реальной работы,
# чтобы показать, где они будут использоваться.
try:
    import torch
    from transformers import AutoTokenizer, AutoModelForSeq2SeqLM # Пример для генерации
    from sentence_transformers import SentenceTransformer # Пример для эмбеддингов
    # from langchain... # Если используете Langchain
    print("Necessary ML/RAG libraries imported (or would be).")
except ImportError as e:
    print(f"Warning: Could not import ML/RAG library: {e}. Faking it.")
    # В реальном проекте здесь должна быть ошибка или установка

def run_research(topic: str) -> str:
    """
    ЗАГЛУШКА: Имитирует выполнение RAG-исследования.
    В реальной версии здесь будет:
    1. Поиск источников (веб, API) по 'topic'.
    2. Загрузка и очистка контента.
    3. Разбиение на чанки.
    4. Получение эмбеддингов чанков (SentenceTransformer).
    5. Сохранение/поиск в векторной БД (FAISS, Chroma).
    6. Поиск релевантных чанков по 'topic'.
    7. Формирование промпта с контекстом (чанками).
    8. Вызов LLM для генерации (Transformers - локально или Inference API).
    9. Возврат сгенерированного текста.
    """
    print(f"[RAG Service Stub] Received topic: '{topic}'")
    print(f"[RAG Service Stub] Pretending to use Hugging Face Token: {'YES' if HUGGINGFACE_TOKEN else 'NO'}")
    print("[RAG Service Stub] Simulating search, retrieval, and generation...")

    # --- Начало блока, где была бы реальная логика ---
    # Пример:
    # search_results = perform_web_search(topic)
    # documents = load_documents(search_results)
    # chunks = split_documents(documents)
    # relevant_chunks = find_relevant_chunks(topic, chunks) # Requires embeddings + vector db
    # prompt = create_prompt(topic, relevant_chunks)
    # generated_text = generate_with_llm(prompt) # Requires generation model
    # --- Конец блока реальной логики ---

    # Имитация задержки
    time.sleep(2)

    # Возвращаем фиктивный результат
    result = f"""
    --- Результат Исследования (ЗАГЛУШКА) ---

    Тема: "{topic}"

    Это демонстрационный ответ. Реальный RAG-пайплайн должен был бы:
    - Найти актуальную информацию в сети.
    - Использовать эмбеддинги (например, из `sentence-transformers`).
    - Найти релевантные фрагменты текста.
    - Сгенерировать ответ с помощью языковой модели (например, из `transformers` с Hugging Face).

    Токен Hugging Face {'обнаружен' if HUGGINGFACE_TOKEN else 'НЕ обнаружен'}.

    (Конец заглушки)
    """
    print("[RAG Service Stub] Finished simulation.")
    return result